requirements/docs.txt
alabaster==0.7.7
Babel==2.2.0
docker-volume==0.1.dev0
docutils==0.12
Jinja2==2.8
lazr.delegates==2.0.3
Mako==1.0.3
MarkupSafe==0.23
nose==1.3.7
Pygments==2.0.2
pytz==2015.7
PyYAML==3.11
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.3.4
sphinx-rtd-theme==0.1.9
zope.component==4.2.2
zope.event==4.1.0
zope.interface==4.1.3
